pyproject.toml
[project]
name = "repl-nix-productattributemanager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "werkzeug>=3.1.3",
    "pillow",
    "flask-migrate>=4.0.7",
]
